Student registration business rules and validation

+++ my-dojo-backend/src/main/java/my/dojo/service/StudentService.java
@@ -27,6 +27,9 @@
 
 package my.dojo.service;
 
+import com.google.common.base.Strings;
+import my.dojo.exception.BusinessException;
+import my.dojo.exception.StudentNotFoundException;
 import my.dojo.model.Student;
 import my.dojo.model.StudentPage;
 import my.dojo.repository.StudentRepository;
@@ -36,19 +39,63 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDate;
+import java.time.Period;
 import java.util.Optional;
 
 @Service
 public class StudentService {
+    private static final int MINIMUM_AGE = 6;
+    private static final int AGE_OF_MAJORITY = 18;
+
     @Autowired
     private StudentRepository studentRepository;
 
-    public Student create(Student student) {
+    public Student create(@NotNull @Valid Student student) throws BusinessException {
+        int age = Period.between(student.getDateOfBirth(), LocalDate.now()).getYears();
+        if (age < MINIMUM_AGE) {
+            throw new BusinessException("Student must have minimum age of 6 years");
+        } else if (age < AGE_OF_MAJORITY) {
+            if (Strings.isNullOrEmpty(student.getGuardianName())) {
+                throw new BusinessException("Guardian name is mandatory for students under 18 years");
+            }
+
+            if (Strings.isNullOrEmpty(student.getGuardianPhone())) {
+                throw new BusinessException("Guardian phone is mandatory for students under 18 years");
+            }
+
+            if (Strings.isNullOrEmpty(student.getGuardianEmail())) {
+                throw new BusinessException("Guardian email is mandatory for students under 18 years");
+            }
+        }
+
+        switch (student.getGradeType()) {
+            case DAN -> {
+                if (student.getGradeRank() < 1 || student.getGradeRank() > 10) {
+                    throw new BusinessException("DAN grades must be between 1st and 10th.");
+                }
+            }
+
+            case KYU -> {
+                if (student.getGradeRank() < 1 || student.getGradeRank() > 7) {
+                    throw new BusinessException("KYU grades must be between 7th and 1st.");
+                }
+            }
+        }
+
+        if (studentRepository.existsByEmail(student.getEmail())) {
+            throw new BusinessException(student.getEmail() + "is already registered.");
+        }
+
         return studentRepository.save(student);
     }
 
-    public Student findById(Long id) {
-        return studentRepository.findById(id).get();
+    public Student findById(Long id) throws StudentNotFoundException {
+        return studentRepository
+                .findById(id)
+                .orElseThrow(() -> new StudentNotFoundException("Student " + id + " not found."));
     }
 
     public StudentPage findAllByNameStartsWith(String name, int page, int size) {
@@ -57,8 +104,10 @@ public class StudentService {
         return new StudentPage(slice.getContent(), slice.getNumber(), slice.getSize(), slice.hasNext());
     }
 
-    public Student update(Long id, Student updatedStudent) {
-        Student student = studentRepository.findById(id).get();
+    public Student update(Long id, Student updatedStudent) throws StudentNotFoundException {
+        Student student = studentRepository
+                .findById(id)
+                .orElseThrow(() -> new StudentNotFoundException("Student " + id + " not found."));
 
         Optional.ofNullable(updatedStudent.getPhoneNumber()).ifPresent(student::setPhoneNumber);
         Optional.ofNullable(updatedStudent.getGradeRank()).ifPresent(student::setGradeRank);
@@ -68,7 +117,10 @@ public class StudentService {
         return studentRepository.save(student);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws StudentNotFoundException {
+        if (!studentRepository.existsById(id)) {
+            throw new StudentNotFoundException("Student " + id + " not found.");
+        }
         studentRepository.deleteById(id);
     }
 }
